refactor(fintual): route Fintual wire code through provider seam (#352)
