+++ src/fintual/fold.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from "vitest"
-import type { GoalPerformanceData } from "./new-performance.ts"
+import type { GoalPerformanceData } from "./provider.ts"
 import { foldGoalPerformanceData } from "./fold.ts"
 
 test("folds reference and recent Goal Performance Data into a Performance Snapshot", () => {
